requirements.txt
pandas
bar_chart_race
